requirements.txt
streamlit==1.32.0
supabase==2.3.0
python-dotenv==1.0.1
openai==1.12.0
langchain==0.1.9
bcrypt==4.1.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
typing-extensions==4.9.0
pydantic==2.6.1
email-validator==2.1.0.post1
aiosmtplib==2.0.2
jinja2==3.1.3
python-json-logger==2.0.7
pytest==8.0.0
pytest-cov==4.1.0
flake8==7.0.0
black==24.1.1
isort==5.13.2
boto3==1.34.34
mixpanel==4.10.0
